+++ raycast/td.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+# @raycast.schemaVersion 1
+# @raycast.title td
+# @raycast.mode compact
+# @raycast.packageName TLTD
+
+# @raycast.argument1 { "type": "text", "placeholder": "task \\\\ description" }
+
+~/.local/bin/td "$1"